requirements.txt
asgiref==3.3.1
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.4
django-comments-dab==2.5.1
django-contrib-comments==1.9.2
djangorestframework==3.12.2
docutils==0.16
gunicorn==20.0.4
pycodestyle==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.2.0
